// src/types/product.ts
export type Product = {
    id: number
    title: string
    price: number
    description: string
    category: string
    image: string
    rating: {
        rate: number
        count: number
    }
}

export type ProductFilters = {
    page?: number
    category?: string
    sort?: 'asc' | 'desc'
}